React Dropdown
React Dropdown example added
(Open and close a dropdown by clicking the container of dropdown;
close by clicking outside the dropdown as well.
If a user clicks the contents of the dropdown, it doesn't close.
It is useful when making smth like nested dropdowns or handling
clicks inside the contents of the dropdown.)